chore: refactor pkg scrapers
* move all run stuff to run.go (delete runscrapers.go)
* move all incremental scrape stuff to incremental.go
